Add basic parser capability
Everything for the parser is contained in the parser directory.
Currently the parser just parses the toy example, but my next task
will be to add command line arguments to specify the project sources
directory.

I also added some stuff to the .gitignore file in the main repo, and
to the .gitignore file in parser. If you want to use IntelliJ Idea,
I would recommend just importing codemap-java-parser through maven,
since maven is what determines the dependencies and such.