Add standalone jar assembly, parsing script
Figured out how to use Maven to make a standalone jar, that comes
with all of the dependencies. Updated the README.md in codemap-java-parser
to reflect this.

Also added the parse.sh script, which will run the parser (and will also
build the jar if it doesn't already exist.) Updated the TODO.md file to
reflect that the parser script is now done.