Add command line options, remove 3rd party method calls
Removes all method calls that are from a 3rd party, using a
minimum threshold of the number of parts of the fully qualified
name of the method that match, starting from the beginning. This
includes calls to native Java libraries.

Added commons-cli as a dependency to provide support for parsing
command line options. Currently there are only two options, or
only one if you don't count the help option, but I think that this
should make our parser more flexible in the future.

Update TODO.md and codemap-java-parser/README.md to reflect some
of these changes.