Refactor code for parsing Java
Made a bunch of static methods that were previously public private,
since main is teh only method that should really be accessed in
ResolveMethodReferences.

Added some more getters and setters for ClassInfo and
MethodDeclarationInfo that make it more convenient to use them instead
of using JSONObject and JSONArray types directly.