Update codemap-java-parser README, more descriptive errors
I found a bug in JavaParser, which I discovered by giving more
descriptive printlns about where the error occurred. I added some
getters for signature and declaration in MethodDeclarationInfo.

Updated the parser README so that it talks about using the `-t` option,
in particular for the toy data.

Updated the parsing/parse.sh script's comment at the top about usage,
because it was incorrect.