Add line numbers to the info gathered from classes and methods
In addition to the other information that is gathered, the parser
now also finds the line numbers that classes and methods are
declared at. This should be useful for going between the vis and
the source code. I'd really like to be able to cmd-click on a method
and have it take me directly to that method.

I also ensure that ~ is expanded to the user's home directory.